InfrastructureModule separated from DomainModule
